+++ src/pages/my-page/setting/manage-account/page.tsx
@@ -1,5 +1,5 @@
 import { usePasswordChangeModal } from "@/features/setting/hooks";
-import { useSettingPermission } from "@/features/setting/lib";
+import { useSettingPermission } from "@/features/setting/hooks";
 import { AccountCancellationModal } from "@/features/setting/ui";
 import { useModal } from "@/shared/lib";
 import { DividerLine } from "@/shared/ui/divider";
